+++ variables.tf
@@ -300,6 +300,71 @@ variable "enable_autoscaling" {
   default     = false
 }
 
+# variable "step_scaling_policies" {
+#   description = "Scaling policies to apply to the scalable target. Supported policy types are StepScaling and TargetTrackingScaling."
+#   type = list(object({
+#     name                            = string
+#     policy_type                     = string
+#     step_scaling_policy_configuration = object({
+#       adjustment_type         = string
+#       cooldown                = number
+#       metric_aggregation_type = string
+#       step_adjustments        = list(object({
+#         metric_interval_lower_bound = number
+#         scaling_adjustment          = number
+#       }))
+#     })
+#   }))
+#   default = []
+# }
+
+variable "step_scaling_policies" {
+  description = "Scaling policies to apply to the scalable target. Supported policy types are StepScaling and TargetTrackingScaling."
+  type = list(object({
+    name                             = string
+    policy_type                      = string
+    step_scaling_policy_configuration = object({
+      adjustment_type         = string
+      cooldown                = number
+      metric_aggregation_type = string
+      step_adjustments        = list(object({
+        metric_interval_lower_bound = optional(number)
+        metric_interval_upper_bound = optional(number)
+        scaling_adjustment          = number
+      }))
+    })
+  }))
+  default = []
+}
+
+variable "target_scaling_policies" {
+  type = list(object({
+    name                            = string
+    policy_type                     = string
+    target_tracking_scaling_policy_configuration = object({
+      target_value       = number
+      scale_in_cooldown  = number
+      scale_out_cooldown = number
+      predefined_metric_specification = object({
+        predefined_metric_type = string
+      })
+    })
+  }))
+  default = []
+}
+
+variable "scheduled_actions" {
+  description = "Scheduled actions to apply to the ecs scalable target."
+  type = list(object({
+    name               = string
+    schedule           = string
+    min_capacity       = number
+    max_capacity       = number
+    timezone           = string
+  }))
+  default = []
+}
+
 variable "max_capacity" {
   description = "(Required) The max capacity of the scalable target."
   type        = number
@@ -327,7 +392,7 @@ variable "scalable_dimension" {
 variable "service_namespace" {
   description = "(Required) The AWS service namespace of the scalable target."
   type        = string
-  default     = ""
+  default     = "ecs"
 }
 
 variable "policy_type" {