+++ CHANGELOG.md
@@ -27,6 +27,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: Add missing aws_ecs_task_definition arguments and showcase them in examples
 - feat: expand volume block of the task definition as it has more configuration
 
+## [1.11.0] - 2024-05-21
+### Changes
+- feat: Add Step Scaling Policy to ECS Service
+- feat: Add Target Tracking Policy to ECS Service
+- feat: Add Scheduled Scaling Policy to ECS Service
+- feat: Add Application Auto Scaling examples
+
 ## [1.10.0] - 2024-05-19
 ### Changes
 - feat: resolve conflict when using service ingress rules with a sg as a source and add support for cidr_blocks and source_security_group_id security group rules groups
@@ -156,8 +163,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: feature update.
 - feat: initial code commit
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.10.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.11.0...HEAD
 
+[1.11.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.11.0
 [1.10.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.10.0
 [1.9.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.9.0
 [1.8.0]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.8.0